refactor services to use DatabaseErrorHandler for error handling
