+++ app/src/main/java/com/baghdad/tudee/data/service/AppConfigurationServiceImpl.kt
@@ -1,13 +1,14 @@
 package com.baghdad.tudee.data.service
 
 import com.baghdad.tudee.data.database.dao.AppConfigurationDao
+import com.baghdad.tudee.data.service.shared.DatabaseErrorHandler
 import com.baghdad.tudee.domain.service.AppConfigurationService
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 
 class AppConfigurationServiceImpl(
     private val appConfigurationDao: AppConfigurationDao
-) : BaseService(), AppConfigurationService {
+) : DatabaseErrorHandler(), AppConfigurationService {
     override suspend fun isDarkTheme(): Flow<Boolean> = executeWithErrorHandling {
         appConfigurationDao.isDarkTheme().map { it == true }
     }
