+++ app/src/main/java/com/baghdad/tudee/data/service/CategoryServiceImpl.kt
@@ -4,6 +4,7 @@ import com.baghdad.tudee.data.database.dao.CategoryDao
 import com.baghdad.tudee.data.mapper.toDto
 import com.baghdad.tudee.data.mapper.toEntities
 import com.baghdad.tudee.data.model.CategoryDto
+import com.baghdad.tudee.data.service.shared.DatabaseErrorHandler
 import com.baghdad.tudee.domain.entity.Category
 import com.baghdad.tudee.domain.service.CategoryService
 import kotlinx.coroutines.flow.Flow
@@ -11,7 +12,7 @@ import kotlinx.coroutines.flow.map
 
 class CategoryServiceImpl(
     private val categoryDao: CategoryDao
-) : CategoryService, BaseService() {
+) : CategoryService, DatabaseErrorHandler() {
     override suspend fun getCategories(): Flow<List<Category>> = executeWithErrorHandling {
         categoryDao.getCategories().map(List<CategoryDto>::toEntities)
     }
