+++ app/src/main/java/com/baghdad/tudee/data/service/TaskServiceImpl.kt
@@ -4,6 +4,7 @@ import com.baghdad.tudee.data.database.dao.TaskDao
 import com.baghdad.tudee.data.mapper.toDto
 import com.baghdad.tudee.data.mapper.toEntities
 import com.baghdad.tudee.data.model.TaskDto
+import com.baghdad.tudee.data.service.shared.DatabaseErrorHandler
 import com.baghdad.tudee.domain.entity.Task
 import com.baghdad.tudee.domain.service.TaskService
 import kotlinx.coroutines.flow.Flow
@@ -12,7 +13,7 @@ import kotlinx.datetime.LocalDate
 
 class TaskServiceImpl(
     private val taskDao: TaskDao
-) : TaskService, BaseService() {
+) : TaskService, DatabaseErrorHandler() {
     override suspend fun getTasksByCategory(categoryId: Long): Flow<List<Task>> =
         executeWithErrorHandling {
             taskDao.getTasksByCategory(categoryId).map(List<TaskDto>::toEntities)
